Fix warnings

+++ src/example_bdds/approximate_adders.cpp
@@ -29,7 +29,7 @@ namespace abo::example_bdds {
             }
         }
 
-        for (int i = partition_size-1;i<num_bits;i++) {
+        for (int i = partition_size-1;i<static_cast<int>(num_bits);i++) {
             BDD temp_carry = mgr.bddZero();
             for (int j = partition_size-2;j >= 0;j--) {
                 if (i-j >= 0) {
@@ -74,13 +74,13 @@ namespace abo::example_bdds {
 
         std::vector<BDD> carry_outs;
         carry_outs.reserve(num_bits);
-        for (int i = 0;i<num_bits;i++) {
+        for (int i = 0;i<static_cast<int>(num_bits);i++) {
             const BDD &a = input1[i];
             const BDD &b = input2[i];
 
             if (i == 0) {
                 carry_outs.push_back(carry_generate(a, b, mgr.bddZero()));
-            } else if (((i+1) % (num_bits / num_partitions) == 0) && (i+1) != num_bits) {
+            } else if (((i+1) % (num_bits / num_partitions) == 0) && (i+1) != static_cast<int>(num_bits)) {
                 BDD temp_carry = mgr.bddZero();
                 for (int j = (partition_size / 2) - 1;j >= 0;j--) {
                     if (i - j >= 0) {
@@ -122,13 +122,13 @@ namespace abo::example_bdds {
 
         std::vector<BDD> carry_outs;
         carry_outs.reserve(num_bits);
-        for (int i = 0;i<num_bits;i++) {
+        for (int i = 0;i<static_cast<int>(num_bits);i++) {
             const BDD &a = input1[i];
             const BDD &b = input2[i];
 
             if (i == 0) {
                 carry_outs.push_back(carry_generate(a, b, mgr.bddZero()));
-            } else if (((i+1) % (num_bits / num_partitions) == 0) && (i+1) != num_bits) {
+            } else if (((i+1) % (num_bits / num_partitions) == 0) && (i+1) != static_cast<int>(num_bits)) {
                 BDD temp_carry = mgr.bddZero();
                 for (int j = prediction_bits - 1;j >= 0;j--) {
                     if (i - j >= 0) {
